requirements.txt
flask_cors==4.0.0
flask==3.0.3
load_dotenv==0.1.0
peewee==3.17.1
pymysql==1.1.0
flask-wtf==1.2.1